Add exercise: Chaining Selectors
This exercise included two HTML elements with one shared class and one
unique class each. The goal was to select these elements individually
using chained class selectors and apply specific styles.
We adjusted their width and height properties. For one, we used a fixed
width and allowed height to scale automatically. For the other, we set
a fixed width and explicitly doubled the height.